refactor: share environment idependent specs (1/2)
Currently there are many duplicated specs for protractor and testbed environment.
Most are equal, some are missplaced, some are slightly different.

Sharing those tests ensures the same tests exists as unit tests and as e2e tests.
Further it reduces code duplication.

Prerequisite #19709 (will need to test equal behavior)
