+++ src/cdk/testing/tests/BUILD.bazel
@@ -30,15 +30,28 @@ ts_library(
 )
 
 ng_test_library(
+    name = "cross_environment_specs",
+    srcs = ["cross-environment.spec.ts"],
+    deps = [
+        ":test_harnesses",
+        "//src/cdk/testing",
+    ],
+)
+
+ng_test_library(
     name = "unit_test_sources",
     srcs = glob(
         [
             "**/*.spec.ts",
             "**/*.spec.d.ts",
         ],
-        exclude = ["**/*.e2e.spec.ts"],
+        exclude = [
+            "cross-environment.spec.ts",
+            "**/*.e2e.spec.ts",
+        ],
     ),
     deps = [
+        ":cross_environment_specs",
         ":test_components",
         ":test_harnesses",
         "//src/cdk/platform",
@@ -57,6 +70,7 @@ ng_e2e_test_library(
         "**/*.spec.d.ts",
     ]),
     deps = [
+        ":cross_environment_specs",
         ":test_harnesses",
         "//src/cdk/testing",
         "//src/cdk/testing/private",
