+++ src/cdk/testing/tests/test-main-component.html
@@ -6,7 +6,7 @@
 <h1 style="height: 100px; width: 200px;">Main Component</h1>
 <div id="username">Hello {{username}} from Angular 2!</div>
 <div class="counters">
-  <button (click)="click()">Up</button><br>
+  <button (click)="click()" id="up-button">Up</button><br>
   <label>Count:</label>
   <div id="counter">{{counter}}</div>
   <label>AsyncCounter:</label>
